+++ qcelemental/models/v1/procedures.py
@@ -67,10 +67,6 @@ class QCInputSpecification(ProtoModel):
         description="Additional information to bundle with the computation. Use for schema development and scratch space.",
     )
 
-    @validator("schema_version", pre=True)
-    def _version_stamp(cls, v):
-        return 1
-
     def convert_v(
         self, target_version: int, /
     ) -> Union["qcelemental.models.v1.QCInputSpecification", "qcelemental.models.v2.AtomicSpecification"]:
@@ -121,10 +117,6 @@ class OptimizationInput(ProtoModel):
             ("molecule_hash", self.initial_molecule.get_hash()[:7]),
         ]
 
-    @validator("schema_version", pre=True)
-    def _version_stamp(cls, v):
-        return 1
-
     def convert_v(
         self, target_version: int, /
     ) -> Union["qcelemental.models.v1.OptimizationInput", "qcelemental.models.v2.OptimizationInput"]:
@@ -136,7 +128,10 @@ class OptimizationInput(ProtoModel):
 
         dself = self.dict()
         if target_version == 2:
+            dself.pop("schema_version")  # changed in v2
             dself.pop("hash_index", None)  # no longer used, so dropped in v2
+
+            dself["initial_molecule"] = self.initial_molecule.convert_v(target_version)
 
             spec = {}
             spec["extras"] = dself.pop("extras")
@@ -160,7 +155,7 @@ class OptimizationResult(OptimizationInput):
     schema_name: constr(  # type: ignore
         strip_whitespace=True, regex=qcschema_optimization_output_default
     ) = qcschema_optimization_output_default
-    schema_version: Literal[1] = 1
+    # Note no schema_version: Literal[1] = Field(1) b/c inherited from OptimizationInput
 
     final_molecule: Optional[Molecule] = Field(..., description="The final molecule of the geometry optimization.")
     trajectory: List[AtomicResult] = Field(
@@ -198,10 +193,6 @@ class OptimizationResult(OptimizationInput):
             raise ValueError(f"Protocol `trajectory:{keep_enum}` is not understood.")
 
         return v
-
-    @validator("schema_version", pre=True)
-    def _version_stamp(cls, v):
-        return 1
 
     def convert_v(
         self,
@@ -242,6 +233,7 @@ class OptimizationResult(OptimizationInput):
 
             dself.pop("hash_index", None)  # no longer used, so dropped in v2
             dself.pop("schema_name")  # changed in v2
+            dself.pop("schema_version")  # changed in v2
 
             v1_input_data = {
                 k: dself.pop(k)
@@ -272,6 +264,7 @@ class OptimizationResult(OptimizationInput):
                 dself["input_data"] = v2_input_data
                 optsubptcl = None
 
+            dself["final_molecule"] = self.final_molecule.convert_v(target_version)
             dself["properties"] = {
                 "return_energy": dself["energies"][-1],
                 "optimization_iterations": len(dself["energies"]),
@@ -318,15 +311,11 @@ class OptimizationSpecification(ProtoModel):
     keywords: Dict[str, Any] = Field({}, description="The optimization specific keywords to be used.")
     protocols: OptimizationProtocols = Field(OptimizationProtocols(), description=str(OptimizationProtocols.__doc__))
 
-    @validator("schema_version", pre=True)
-    def _version_stamp(cls, v):
-        return 1
-
     @validator("procedure")
     def _check_procedure(cls, v):
         return v.lower()
 
-    # NOTE: def convert_v() is missing deliberately. Because the v1 schema has a minor and different role only for
+    # NOTE: def convert_v() is missing deliberately. Because the v1 schema has a different role only for
     #   TorsionDrive, it doesn't have nearly enough info to create a v2 schema.
 
 
@@ -397,10 +386,6 @@ class TorsionDriveInput(ProtoModel):
         assert value.driver == DriverEnum.gradient, "driver must be set to gradient"
         return value
 
-    @validator("schema_version", pre=True)
-    def _version_stamp(cls, v):
-        return 1
-
     def convert_v(
         self, target_version: int, /
     ) -> Union["qcelemental.models.v1.TorsionDriveInput", "qcelemental.models.v2.TorsionDriveInput"]:
@@ -436,7 +421,8 @@ class TorsionDriveInput(ProtoModel):
 
             dtop = {}
             dtop["provenance"] = dself.pop("provenance")
-            dtop["initial_molecules"] = dself.pop("initial_molecule")
+            dself.pop("initial_molecule")
+            dtop["initial_molecules"] = [mol.convert_v(target_version) for mol in self.initial_molecule]
             dtop["specification"] = tdspec
             dself.pop("schema_name")
             dself.pop("schema_version")
@@ -458,7 +444,7 @@ class TorsionDriveResult(TorsionDriveInput):
     """
 
     schema_name: constr(strip_whitespace=True, regex=qcschema_torsion_drive_output_default) = qcschema_torsion_drive_output_default  # type: ignore
-    schema_version: Literal[1] = 1
+    # Note no schema_version: Literal[1] = Field(1) b/c inherited from TorsionDriveInput
 
     final_energies: Dict[str, float] = Field(
         ..., description="The final energy at each angle of the TorsionDrive scan."
@@ -480,10 +466,6 @@ class TorsionDriveResult(TorsionDriveInput):
     )
     error: Optional[ComputeError] = Field(None, description=str(ComputeError.__doc__))
     provenance: Provenance = Field(..., description=str(Provenance.__doc__))
-
-    @validator("schema_version", pre=True)
-    def _version_stamp(cls, v):
-        return 1
 
     def convert_v(
         self, target_version: int, /, *, external_input_data: "TorsionDriveInput" = None
@@ -536,7 +518,8 @@ class TorsionDriveResult(TorsionDriveInput):
             dtop["stderr"] = dself.pop("stderr")
             dtop["success"] = dself.pop("success")
             dtop["final_energies"] = dself.pop("final_energies")
-            dtop["final_molecules"] = dself.pop("final_molecules")
+            dself.pop("final_molecules")
+            dtop["final_molecules"] = {k: m.convert_v(target_version) for k, m in self.final_molecules.items()}
             dtop["optimization_history"] = {
                 k: [opthist_class(**res).convert_v(target_version) for res in lst]
                 for k, lst in dself["optimization_history"].items()